Add clade mitochondrial sequence extractor
extract :: Taxid -> ProSeq -> ProSeq

This function takes a Taxon id for a high level clade (usually, though
it could be a single species) and collects the taxon ids for all species
beneath it on the NCBI common tree. Next it maps these taxon ids to
scientific names. Finally it reads a Fasta file, where headers have the
form

```
>id desc [Species Name]
```

and extracts all entries where the species name matches one of those on
the list.